// client/src/type/EventResponse.ts
export enum EventResponseStatus {
  ROOM_ID = "ROOM_ID",
  ROOM_STATE_CHANGED = "ROOM_STATE_CHANGED",
  GAME_STATE_CHANGED = "GAME_STATE_CHANGED",
  IN_MEMORY_AUTH = "IN_MEMORY_AUTH",
  SPOTIFY_AUTH = "SPOTIFY_AUTH",
  SPOTIFY_ACCESS_TOKEN = "SPOTIFY_ACCESS_TOKEN",
  UNAUTHORIZED_ACCESS_TOKEN = "UNAUTHORIZED_ACCESS_TOKEN",
  SEARCH_PLAYLISTS = "SEARCH_PLAYLISTS",
  CREATE_ROOM = "CREATE_ROOM",
  JOIN_ROOM = "JOIN_ROOM",
  LEAVE_ROOM = "LEAVE_ROOM",
  CHANGE_PLAYER_NAME = "CHANGE_PLAYER_NAME",
  CHANGE_PLAYER_COLOR = "CHANGE_PLAYER_COLOR",
  REMOVE_PLAYER = "REMOVE_PLAYER",
  ADD_PLAYLIST = "ADD_PLAYLIST",
  REMOVE_PLAYLIST = "REMOVE_PLAYLIST",
  NEXT_TURN = "NEXT_TURN",
  ADD_CURRENT_CARD = "ADD_CURRENT_CARD",
  REMOVE_CURRENT_CARD = "REMOVE_CURRENT_CARD",
  RETURN_CURRENT_CARD = "RETURN_CURRENT_CARD",
  MOVE_CURRENT_CARD = "MOVE_CURRENT_CARD",
  ADD_TOKEN = "ADD_TOKEN",
  REMOVE_TOKEN = "REMOVE_TOKEN",
  START_GAME = "START_GAME",
  INVALID_GAME_STATUS = "INVALID_GAME_STATUS",
  INVALID_PLAYER_ID_FOR_TOKEN_ID = "INVALID_PLAYER_ID_FOR_TOKEN_ID",
  PLAYER_HOST_CANNOT_BE_REMOVED = "PLAYER_HOST_CANNOT_BE_REMOVED",
  PLAYER_HOST_MUST_START_GAME = "PLAYER_HOST_MUST_START_GAME",
  PLAYER_NAME_ALREADY_EXISTS = "PLAYER_NAME_ALREADY_EXISTS",
  PLAYER_COLOR_ALREADY_EXISTS = "PLAYER_COLOR_ALREADY_EXISTS",
  PLAYER_ALREADY_IN_ROOM = "PLAYER_ALREADY_IN_ROOM",
  PLAYER_NAME_NOT_SET = "PLAYER_NAME_NOT_SET",
  NO_PLAYLIST_SELECTED = "NO_PLAYLIST_SELECTED",
  PLAYLIST_ALREADY_IN_ROOM = "PLAYLIST_ALREADY_IN_ROOM",
  GAME_NOT_FOUND = "GAME_NOT_FOUND",
  PLAYER_NOT_FOUND = "PLAYER_NOT_FOUND",
  CONNECTION_NOT_FOUND = "CONNECTION_NOT_FOUND",
  PLAYLIST_NOT_FOUND = "PLAYLIST_NOT_FOUND",
  CARD_MOVEMENT_NOT_FOUND = "CARD_MOVEMENT_NOT_FOUND",
  TOKEN_MOVEMENT_NOT_FOUND = "TOKEN_MOVEMENT_NOT_FOUND",
  AUTH_TYPE_NOT_FOUND = "AUTH_TYPE_NOT_FOUND",
  SPOTIFY_API_GET_PLAYLIST_ITEMS = "SPOTIFY_API_GET_PLAYLIST_ITEMS",
  INVALID_PILE = "INVALID_PILE",
  CARD_ALREADY_IN_PILE = "CARD_ALREADY_IN_PILE",
  CARD_ALREADY_IN_CURRENT_DECK = "CARD_ALREADY_IN_CURRENT_DECK",
  CARD_NOT_IN_CURRENT_DECK = "CARD_NOT_IN_CURRENT_DECK",
  TOKEN_ALREADY_IN_CURRENT_DECK = "TOKEN_ALREADY_IN_CURRENT_DECK",
  TOKEN_ALREADY_IN_PLAYER_DECK = "TOKEN_ALREADY_IN_PLAYER_DECK",
}

export type EventResponse<T> = {
  success: boolean;
  status: EventResponseStatus;
  code: number;
  data?: T;
  message?: string;
}